Added Prog1 Code
Initial Commit for Prog1 Code. So far I have written code for validating
arguments and parsing the input file. I have also started the brute
force algorithm for the Convex Hull of the input points, but a lot of
work remains for it (namely unit tests, clean up, and a solid decision
on the function inputs/returns).