Updates to Prog1
This commit does a few things:

- Adds support for the -debug flag [although debug output is not
properly implemented yet]

- Adds the performance timing for the calculation of the convex hull

- Did some cleanup to calculateConvexHull

I also found a bug in calculateConvexHull. The algorithm currently does
not process colinear points in the set properly. This appears to be due
to an issue with Line2D.relativeCCW [which in my opinion is broken].
Basically, it will not return a proper value for a point which is
colinear to the line, but does not reside between the two points which
define the line. This however should be easy to fix.