Prog1: Fix calculateConvexHull
In order to fix the broken Line2d.relativeCCW function, we change the
function to use two separate calls to Line2D.ptSegDist and
Line2D.ptLineDist. If ptLineDist is 0, and ptSegDist is greater than 0,
then the point we are inspecting is colinear but outside of the range of
the points defining our line.

This new code allows us to ignore potential line segments that can be
replaced by a single larger line segment as well, which means we should
always fine the minimum number of lines that make up the convex hull,
without any redundant lines added into the mix.