Fixed a bug in QuickHull
- I found a bug in my QuickHull algorithm related to relativeCCW (what a
surprise). The output of relativeCCW is dependent on the first point
passed in to the Line2D constructor; if you want to use it and maintain
the idea of a normally oriented x-y coordinate axis, you need to be sure
that the first point P1 that is passed into the constructor is the
closest point to the origin.

- Also added unmodified Analysis and Readme files